requirements.txt
hyperliquid
pandas
numpy
websockets
asyncio
python-dotenv